requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.0
pandas>=2.1.0
openpyxl>=3.1.2

# API integrations
hunter>=2.0.1
clearbit>=0.1.7

# Email
smtplib2>=0.2.1
sendgrid>=6.10.0
email-validator>=2.1.0

# Storage
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.11.3
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.7

# Web framework
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.1
wtforms>=3.1.0
gunicorn>=21.2.0

# Web scraping
playwright>=1.40.0

# Scheduling
schedule>=1.2.0
celery>=5.3.4
redis>=5.0.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.2
validators>=0.22.0

# Development
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.10.1
flake8>=6.1.0
mypy>=1.7.0

# Web scraping utilities
webdriver-manager>=4.0.1
fake-useragent>=1.4.0
retrying>=1.3.4

# Data processing
numpy>=1.25.2
matplotlib>=3.8.0
seaborn>=0.13.0

# Logging and monitoring
loguru>=0.7.2
sentry-sdk>=1.38.0
